PR Feedback; Refactor on TransferList.tsx; Creating TransferListColumn.tsx;
